+++ src/lib/models/reflections/index.ts
@@ -16,3 +16,4 @@ export { SignatureReflection } from "./signature";
 export { TypeParameterReflection, VarianceModifier } from "./type-parameter";
 export { splitUnquotedString } from "./utils";
 export type { ReflectionVariant } from "./variant";
+export { ReflectionSymbolId, type ReflectionSymbolIdString } from "./id";